refactor(shared): extract tool execution helpers into shared.tool_execution
Tool result processing and terminal tool normalization are compatibility shims across Open WebUI versions. Centralize them so all plugins handle result shapes the same way.
