+++ src/owui_ext/tools/llm_review.py
@@ -35,7 +35,10 @@ from owui_ext.shared.prompt_utils import (
 )
 from owui_ext.shared.terminal_events import emit_terminal_tool_event
 from owui_ext.shared.tool_event_metadata import CITATION_TOOLS, TERMINAL_EVENT_TOOLS
-from owui_ext.shared.users import _normalize_user
+from owui_ext.shared.tool_execution import (
+    normalize_terminal_tools_result,
+    process_tool_result,
+)
 
 try:
     import markdown as _markdown_mod
@@ -43,8 +46,6 @@ except ImportError:
     _markdown_mod = None
 
 log = logging.getLogger(__name__)
-
-_core_process_tool_result = None
 
 
 # ============================================================================
@@ -2948,50 +2949,6 @@ async def cleanup_mcp_clients(mcp_clients: dict) -> None:
             log.debug(f"[LLMReview] Error cleaning up MCP client: {e}")
 
 
-async def process_tool_result(
-    *,
-    tool_function_name: str = "tool",
-    tool_type: str,
-    tool_result: Any,
-    direct_tool: bool = False,
-    request: Optional[Request] = None,
-    metadata: Optional[dict] = None,
-    user: Any = None,
-) -> tuple[Any, list, list]:
-    """Process tool result into (payload, files, embeds) using core when available."""
-    global _core_process_tool_result
-    if _core_process_tool_result is None:
-        try:
-            from open_webui.utils.middleware import process_tool_result as fn
-
-            if fn is not None:
-                _core_process_tool_result = fn
-        except ImportError:
-            pass
-    if _core_process_tool_result is not None:
-        return await maybe_await(_core_process_tool_result(
-            request,
-            tool_function_name,
-            tool_result,
-            tool_type,
-            direct_tool=direct_tool,
-            metadata=metadata if isinstance(metadata, dict) else {},
-            user=_normalize_user(user),
-        ))
-    # Fallback for Open WebUI < 0.8.x
-    if isinstance(tool_result, tuple):
-        tool_result = tool_result[0] if tool_result else ""
-    elif direct_tool and isinstance(tool_result, list) and len(tool_result) == 2:
-        tool_result = tool_result[0]
-    if isinstance(tool_result, (dict, list)):
-        tool_result = json.dumps(tool_result, indent=2, ensure_ascii=False)
-    elif tool_result is not None and not isinstance(tool_result, str):
-        tool_result = str(tool_result)
-    return tool_result, [], []
-
-
-
-
 def parse_model_ids(value: Any) -> list[str]:
     """Parse a comma-separated string or list of model IDs.
 
@@ -3019,33 +2976,6 @@ def parse_model_ids(value: Any) -> list[str]:
             continue
         result.append(text)
     return result
-
-
-def normalize_terminal_tools_result(*, terminal_tools_result: Any, extra_params: Optional[dict]) -> dict:
-    """Normalize get_terminal_tools() return value across Open WebUI versions."""
-    terminal_system_prompt = None
-    terminal_tools = terminal_tools_result
-
-    if (
-        isinstance(terminal_tools_result, tuple)
-        and len(terminal_tools_result) == 2
-        and isinstance(terminal_tools_result[0], dict)
-    ):
-        terminal_tools = terminal_tools_result[0]
-        if isinstance(terminal_tools_result[1], str):
-            stripped_prompt = terminal_tools_result[1].strip()
-            if stripped_prompt:
-                terminal_system_prompt = stripped_prompt
-
-    if isinstance(extra_params, dict):
-        if terminal_system_prompt:
-            extra_params["__terminal_system_prompt__"] = terminal_system_prompt
-        else:
-            extra_params.pop("__terminal_system_prompt__", None)
-
-    if isinstance(terminal_tools, dict):
-        return terminal_tools
-    return {}
 
 
 # ============================================================================
