+++ src/owui_ext/tools/parallel_tools.py
@@ -29,7 +29,10 @@ from pydantic import BaseModel, Field
 from owui_ext.shared.async_utils import maybe_await
 from owui_ext.shared.terminal_events import emit_terminal_tool_event
 from owui_ext.shared.tool_event_metadata import CITATION_TOOLS, TERMINAL_EVENT_TOOLS
-from owui_ext.shared.users import _normalize_user
+from owui_ext.shared.tool_execution import (
+    normalize_terminal_tools_result,
+    process_tool_result,
+)
 
 log = logging.getLogger(__name__)
 
@@ -39,9 +42,6 @@ class ToolCallItem(BaseModel):
 
     name: str = Field(description="Tool function name to call")
     args: dict = Field(default_factory=dict, description="Arguments to pass to the tool")
-
-
-_core_process_tool_result = None
 
 
 # ============================================================================
@@ -200,75 +200,6 @@ async def execute_direct_tool_call(
             },
         }
     )
-
-
-def normalize_terminal_tools_result(*, terminal_tools_result: Any, extra_params: Optional[dict]) -> dict:
-    """Normalize get_terminal_tools() return value across Open WebUI versions."""
-    terminal_system_prompt = None
-    terminal_tools = terminal_tools_result
-
-    if (
-        isinstance(terminal_tools_result, tuple)
-        and len(terminal_tools_result) == 2
-        and isinstance(terminal_tools_result[0], dict)
-    ):
-        terminal_tools = terminal_tools_result[0]
-        if isinstance(terminal_tools_result[1], str):
-            stripped_prompt = terminal_tools_result[1].strip()
-            if stripped_prompt:
-                terminal_system_prompt = stripped_prompt
-
-    if isinstance(extra_params, dict):
-        if terminal_system_prompt:
-            extra_params["__terminal_system_prompt__"] = terminal_system_prompt
-        else:
-            extra_params.pop("__terminal_system_prompt__", None)
-
-    if isinstance(terminal_tools, dict):
-        return terminal_tools
-    return {}
-
-
-async def process_tool_result(
-    *,
-    tool_function_name: str = "tool",
-    tool_type: str,
-    tool_result: Any,
-    direct_tool: bool = False,
-    request: Optional[Request] = None,
-    metadata: Optional[dict] = None,
-    user: Any = None,
-) -> tuple[Any, list, list]:
-    """Process tool result into (payload, files, embeds) using core when available."""
-    global _core_process_tool_result
-    if _core_process_tool_result is None:
-        try:
-            from open_webui.utils.middleware import process_tool_result as fn
-
-            if fn is not None:
-                _core_process_tool_result = fn
-        except ImportError:
-            pass
-    if _core_process_tool_result is not None:
-        return await maybe_await(_core_process_tool_result(
-            request,
-            tool_function_name,
-            tool_result,
-            tool_type,
-            direct_tool=direct_tool,
-            metadata=metadata if isinstance(metadata, dict) else {},
-            user=_normalize_user(user),
-        ))
-    # Fallback for Open WebUI < 0.8.x
-    if isinstance(tool_result, tuple):
-        tool_result = tool_result[0] if tool_result else ""
-    elif direct_tool and isinstance(tool_result, list) and len(tool_result) == 2:
-        tool_result = tool_result[0]
-    if isinstance(tool_result, (dict, list)):
-        tool_result = json.dumps(tool_result, indent=2, ensure_ascii=False)
-    elif tool_result is not None and not isinstance(tool_result, str):
-        tool_result = str(tool_result)
-    return tool_result, [], []
 
 
 async def execute_single_tool(
